Move components out of ui folder
