+++ src/components/InventoryItems.tsx
