+++ src/components/hooks/useGetDefaultInventoryItem.ts
